+++ src/equisolve/numpy/scripts/soap.py
@@ -1,0 +1,1 @@
+# TODO
